+++ CHANGES.rst
@@ -1,7 +1,10 @@
 0.2 (unreleased)
 ================
 
-* Nothing changed yet.
+* Changed the ``ticket.changed`` channel so that the ``"new_values"``
+  keyword contains all of the changed ticket's fields, including those that
+  did not change.  This allows more flexibility in formatting the ticket
+  changed messages.
 
 
 0.1 (03/02/2018)